Make ruff and mypy clean across the package
ruff goes 793 -> 0 and mypy 218 -> 0 ("Success: no issues found"), so the whole
`make check` gate (ruff + black + mypy) is green alongside the pre-commit hooks.

Mechanical ruff fixes: import sorting (I001), unused imports (F401), unused
noqa (RUF100), set/dict comprehensions (C4xx), unused unpacked variables
(RUF059, renamed to _), unused loop variables (B007), loop-variable rebinding
(PLW2901, renamed to *_raw), collapsible ifs (SIM102/108), long lines (E501,
wrapped comments/docstrings/f-strings), and ambiguous unicode in comments and
docstrings (RUF002/003, en-dash/minus/times -> ASCII).

Genuine defects the linters surfaced, fixed rather than silenced:
- F811: three duplicated test definitions in test_manybody_basis.py meant the
  earlier copy silently never ran. Removed the true duplicates and renamed the
  serial test_state_mpi (comm=None, no mpi marker) to test_state.
- F821: mpi_comm.py referenced ManyBodyBlockState from annotations without
  importing it.
- F841: eigensolvers.py computed max_subspace_blocks (and num_wanted) and never
  passed them to any solver -- dead since a refactor.
- A001/A002: op_parser.py shadowed the `str` builtin as a parameter it also
  rebound; renamed to `text`.

Type fixes for mypy: accurate Optional/union signatures, annotations for
otherwise-uninferable containers, @overload for Basis.contains (its return type
depends on whether it is handed one state or many), Iterator (not Iterable)
where next() is called, and explicit None-narrowing asserts in the MPI paths of
basis_split.py/gf_units.py where the invariants hold but are not visible to the
checker. manybody_basis.py drops a dead pre-3.10 collections fallback.

greens_function.py keeps its backward-compat re-export block, now marked
`# noqa: F401`: those names are imported purely so `greens_function.X` keeps
resolving after the gf_primitives/gf_convergence/gf_shift_recycling split, and
an unannotated `ruff --fix` strips them, which breaks nine test modules at
import time.

Test gate green at serial, -n 2 and -n 3 (the split path).

+++ src/impurityModel/test/test_sectorization.py
@@ -133,8 +133,8 @@ def test_restrictions_from_charges_and_occupations():
 # ---------------------------------------------------------------------------
 
 import impurityModel.ed.product_state_representation as psr  # noqa: E402
+from impurityModel.ed.basis_transcription import build_dense_matrix  # noqa: E402
 from impurityModel.ed.manybody_basis import Basis  # noqa: E402
-from impurityModel.ed.basis_transcription import build_dense_matrix
 
 _BASIS_KW = dict(impurity_orbitals={0: [[0, 1, 2, 3]]}, bath_states=({0: [[]]}, {0: [[]]}), verbose=False)
 
@@ -183,8 +183,7 @@ def test_automatic_restrictions():
     # each <N_S> is integer for the non-degenerate GS).
     local = [bytes(d.to_bytearray()) for d in full.local_basis]
     gs_occ = [
-        int(round(sum(abs(gs[i]) ** 2 * len(charge & _occset(local[i])) for i in range(len(local)))))
-        for charge in charges
+        round(sum(abs(gs[i]) ** 2 * len(charge & _occset(local[i])) for i in range(len(local)))) for charge in charges
     ]
     assert gs_occ == [1, 1]  # half-filled singlet sector
 
@@ -291,7 +290,7 @@ def test_hf_seed_finds_ground_state_sector():
     # The mean-field seed lands on the correct integer impurity occupation of the correlated GS,
     # and says so: an unconverged seed is not an occupation and callers must reject it.
     assert converged
-    assert winning_N0 == {0: int(round(exact_imp_occ))}
+    assert winning_N0 == {0: round(exact_imp_occ)}
 
 
 def test_unconverged_hf_seed_is_rejected(monkeypatch):
